sell item inventory

+++ Assets/Script/Inventory/Item.cs
@@ -7,7 +7,7 @@ public class Item : ScriptableObject
 {
     public string itemId;
     public string itemName;
-    public int itemCoin;
+    public float itemCoin;
 
     public Sprite icon;
 }
